tests(motion_planner): split up test_motion_planner into ompl / simple
